requirements.txt
attrs==21.4.0
beautifulsoup4==4.10.0
blinker==1.4
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-Dance==5.1.0
Flask-JWT-Extended==4.3.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
lxml==4.8.0
MarkupSafe==2.1.0
numpy==1.22.3
oauthlib==3.2.0
pandas==1.4.1
pickleshare==0.7.5
PyJWT==2.3.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.32
threadpoolctl==3.1.0
tmdbv3api==1.7.6
urllib3==1.26.8
URLObject==2.4.3
Werkzeug==2.0.3
